Created shop UI
Created the shop UI and system that would allow a more easier time to add and store items into the shop, Also have a player script to store vaules on the player and have an interface that can interact with other scripts to allow access between both the stored vaules and other scripts. do not have a button interaction yet since we do not have items to activate so it remains in pending. the shop desgin is WIP